Refactor hexCell components
